fix: make note parsing more flexible
Although a note w/o any content isn't especially useful, it is still
valid according to Ledger. Moreover, supporting such notes fixes glitchy
syntax highlighting when comments are being added: without this fix,
adding a note to a transaction or posting would cause the next posting
to be incorrectly parsed as a comment.

The change for "no leading whitespace" is basically the same, except
that's just stylistic: ledger doesn't care if a note has leading
whitespace, so perfectly valid notes can cause parse errors, which leads
to the same issue as above: the next posting is incorrectly highlighted
as a comment.