fix: allow special chars in payees
Bank statements often include asterisks in the transaction name, and Ledger
seems to place few limits on what characters can make up a payee name, but this
grammar had been excluding ! and *. This removes that restriction so that
payees can now contain ! and *.

Payees still cannot *start* with !, * or (, presumably to disambiguate with
codes and statuses. Interestingly, Ledger doesn't seem to share this
limitation, and a payee can start start with a !, * or even (, so long as
that payee is preceeded by a code. For example, Ledger doesn't seem to mind
any of these:

2024-11-18 () ( Payee)
2024-11-18 () ! Payee
2024-11-18 () * Payee

"Interestingly" because these are obviously problematic names (removing the
empty code would cause them all to mean something else), not because I think we
should consider supporting them. :)