fix: notes following payee
Notes that follow the payee are incorrectly being consumed as part of the
payee. There is a test that includes such a note, but the error was missed
because the assertion only included 1 note node, not 2 as there should have
been all along.

This change correctly parses these notes separate from the payee, includes a
test to make to make future regressions more obvious, and corrects the existing
test to expect 2 note nodes.