fix: correctly parse amounts without commodities followed by other elements
When a posting amount has no commodity, and then it's followed by a space and
pretty much anything else, the parser behaves _almost_ correctly except that it
injects empty error nodes into the tree. I ran into this when working on #34.

In short, given a transaction like this:

```
2020/12/02
    Account  2 @ $1
```

The parser would recognize the first amount (2) and also the price (`@ $1`), but
it would also show an `(ERROR)` node for the space in `2 @`. The same would
happen for lot prices (eg `2 {$1}`), balance assertions (eg `2 = $1`) and even
comments (eg `2 ;foo`).

The parse became incorrect if a second space was added between the amount and
the syntax char (eg `2  @ $1`) or if the comment char was followed by a space
(eg `2 ; foo`). For the former, there would be an error indicating a missing
commodity token. For the latter, `;` would be an error, and `foo` would be
parsed as the commodity. (!?)

Poking around the grammar, I tracked the issue down to the "amount followed by
commodity" rule, and I was eventually able to fix this by removing the "amount
without commodity" variant and changing the "amount followed by commodity" into
"amount followed by _optional_ commodity".